+++ src/main/java/ru/nemodev/project/quotes/dao/quote/QuoteDAO.java
@@ -14,7 +14,7 @@ public interface QuoteDAO
      * @param count
      * @return
      */
-    List<Quote> getRandom(Long count);
+    List<Quote> getRandom(Integer count);
 
     /**
      *
